完成mconfig sdk config 编写 简单调试达到最终效果
